Update spy to 1.0.5
Looks like they fixed a bug where the class was getting passed as an
argument to a stub block, so I had to update my janky test.

Also kwargs are first class now so another test needed updating

Also loosened the rescue in ActivityNotifier so it doesn't swallow
things like ArgumentError which it was doing here.