894 add complete button to appointment page (#1328)
# What it does

Adding a complete button to the appointment details page. Clicking this
button completes the appointment and returns to the appointment list
page. If viewing a completed appointment, the button will restore the
appointment and leave the user on the appointment details page.

# Why it is important

Users have requested a button here to avoid having to go back to the
appointment list page before completing the appointment.

# UI Change Screenshot

Before:
![Screenshot 2024-02-11
115409](https://github.com/chicago-tool-library/circulate/assets/2738059/18863681-a610-4c85-868c-e54987f2882a)

After:
![Screenshot 2024-02-11
115110](https://github.com/chicago-tool-library/circulate/assets/2738059/9809be4e-e874-4105-9380-62e3e6d9bd32)

# Implementation notes

Not 100% sure about the creation of the new controller and partial view
approach. If the reviewer could make sure that it is following Ruby on
Rails best practices, that would be great!

# Your bandwidth for additional changes to this PR

I have the time and interest to make additional changes to this PR based
on feedback.