Add ItemPools and update reservations accordingly. (#1327)
# What it does

Adds the concept of an item pool and item pile (😎) to the app. This
allows groups members to reserve a quantity of an item pool ("5 cordless
drills"), and also to have item pools that don't have uniquely numbered
items (e.g. work gloves).

# Why it is important

This is the next stage of building out group lending. Once this is
merged, the main data models to represent items and reservations will be
in place. Next up is how we handle pickups and actually checking out
items.

# Implementation notes

* Some of the database querying code isn't well factored yet.

# Your bandwidth for additional changes to this PR

_Please choose one of the following to help the project maintainers
provide the appropriate level of support:_

- [x] I have the time and interest to make additional changes to this PR
based on feedback.