Add items to pickups for group reservations (#1388)
# What it does

Extends the "pickup" page to allow for specific items to be added as
they are pulled from the shelf and added to the order.

# UI Change Screenshot

![2024-03-10 19 03
15](https://github.com/chicago-tool-library/circulate/assets/3331/54a75b6d-56b1-4291-84a1-43b4e6249232)

# Implementation notes

* `ReservationLoan` objects belong to either a `ReservableItem` for
uniquely numbered items (item pools) or an `ItemPool` for non-uniquely
numbered items (item pile). I can be convinced that there is a better
way to model this.
* The use of `autofocus` to keep the item ID input accepting keyboard
input is a temporary solution. We'll likely want something slightly more
involved eventually. It's a bad practice from an accessibility
standpoint; however, the main use case of the page is someone who has a
barcode scanner and is going to scan **a lot** of items. Our current
system sometimes loses focus and folks will scan for a while without
realizing that it's not doing anything. So having some kind of focus
control here is necessary.
* I am convinced we should rename `Pickup` to `Fulfillment`, but I don't
want to do that in this PR as there are already a lot of changes.