Remove categories from appointments list (#1405)
# What it does

As discussed with librarians in Slack, categories aren't all that useful
to have in the appointments list. The categories aren't currently
indicative of location and because some of them are long they add a lot
of vertical space to the list.

# Why it is important

We're working on making the appointments page more useful to librarians
when they prepare orders, and their feedback is important!

# UI Change Screenshot

Before:

![2024-03-15 at 13 26
48@2x](https://github.com/chicago-tool-library/circulate/assets/37534/bacf89a7-cb2a-475f-8e79-59bd7d27307a)

After:

![2024-03-15 at 13 27
02@2x](https://github.com/chicago-tool-library/circulate/assets/37534/af9f0f50-4b4b-4805-9949-47173f6a5b3d)

# Your bandwidth for additional changes to this PR

_Please choose one of the following to help the project maintainers
provide the appropriate level of support:_

- [x] I have the time and interest to make additional changes to this PR
based on feedback.