Convert new appointments page to feature flag
In our weekly meeting yesterday we reviewed the staff feedback on the
new appointments page, and we decided the next best step is to pull the
old page back and place the new look behind a feature flag to let us
iterate on it.

This feature flag is able to be set with an env var but also able to be
overridden by a query param to allow easy toggling back and forth
between old and new looks.