Specify node/yarn version where heroku buildpack will find them (#1432)
# What it does

As a part of #1185 we need to tell the nodejs buildpack which version of
node to deploy. To do this we follow [the heroku
docs](https://devcenter.heroku.com/articles/nodejs-support#specifying-a-node-js-version)
and add to package.json.

Added a few unit tests to help us remember to keep versions consistent
in the various config files.

# Why it is important

Per #1185 this gets us in line with Heroku's recommendations and quells
a big warning in the logs every time we deploy.