Refactor Google Calendar authentication (#1444)
# What it does

* Use
[googleauth](https://github.com/googleapis/google-auth-library-ruby)
library to handle token generation (we previously were doing this
ourselves)
* Allows the use of a GCP service account to authenticate (we'll use
this in production)
* Adds documentation that explains how to setup authentication in
development for anyone wanting to work on this code.
* This should fix #1437.

# Why it is important

* Google made a change to how authentication was working in production,
and as a part of sorting that out, I wanted to move toward using a
service account instead of OAuth. This turned out to be tricky to setup
for local development (there are a lot of things to turn on/grant in a
brand new GCP project, which is what most folks will start with), so for
that use case I've preserved the existing OAuth option.

# Implementation notes

* We have a volunteer shift scheduling system that is currently not used
by CTL. There's a decent amount of code for that functionality that we
could consider removing, which would let us really simplify the Google
Calendar parts of the codebase.
* The use of `.profile` to write the contents of an environmental
variable to a file feels a little gross, but it seems to be a common way
to work around the fact that Heroku's 12-factor philosophy and Google's
default way of handling credentials are at odds.

# Your bandwidth for additional changes to this PR

_Please choose one of the following to help the project maintainers
provide the appropriate level of support:_

- [x] I have the time and interest to make additional changes to this PR
based on feedback.

# Deployment Plan

* I've already setup the environment variables in production. I will
create some new calendars for staging and configure a service account
there as well so that we can verify things are working before going to
prod.