Wait for logs to write to appsignal before exiting rake tasks. (#1445)
# What it does

I noticed that logs for tasks run through the Heroku scheduler aren't
showing up in AppSignal. [Their docs mention needing to add some code to
ensure that logs are send before the process
exits](https://docs.appsignal.com/ruby/integrations/rake.html#rake-tasks-and-containers),
so I added that to the calendar tasks. If it works, we can add it to the
other ones that send emails and the like.

# Why it is important

It's nice to keep an eye on if these tasks are running and how many
events they are updating.

# Your bandwidth for additional changes to this PR

- [x] I have the time and interest to make additional changes to this PR
based on feedback.