Move appointments between pending/completed lists (#1450)
# What it does

Fixes #1438 by moving appointments to the appropriate list when the are
completed or restored. This feature was originally implemented in
dbf9983 but removed in #1181 as a part of the upgrade from Turbolinks to
Turbo.

# Why it is important

Staff have given us feedback that they miss having the app automatically
move completed appointments to the list at the bottom of the page
without needing a page refresh.

# Implementation notes

* The new implementation works on both the new and old UI for the
appointment list.
* `FEATURE_NEW_APPOINTMENTS_PAGE` now defaults to `false` in the test
environment (so we can test that the current default is working in
tests). We could also write tests for both, but that can be a followup.
* I also fixed a bug where completion didn't work properly when
`new=true` using a query param.

---------

Co-authored-by: Paul Hinze <phinze@phinze.com>