Add Ahoy for tracking and begin recording events for tool search (#1454)
# What it does

This installs the [`ahoy_matey`](https://github.com/ankane/ahoy) gem for
first-party analytics and begins using it to track when members do a
search, view an item, and place a hold.

# Why it is important

As we aim to improve search in the system (#1240), it will be helpful to
see what users are actually searching for and whether they are finding
it.

# Implementation notes

* 🤫 By using a first-party analytics system (i.e. one that stores data
in the same database we already use and control) we're not storing or
exposing any more user information than we already have via our member
information and logs.
* 🏋🏻 It's worth noting that this will put more storage pressure on the
database, because it's recording rows for stuff we expect users to do a
lot. So we'll want to keep an eye on database growth.
* ♻️ Along those lines, will want to add [a job that prunes old
data](https://github.com/ankane/ahoy?tab=readme-ov-file#data-retention)
but I figure that can come in a follow-on PR.
* 📈 We can add as many more events as we like over time, but I figure
starting relatively simple is the way to go.