Added pagination to account holds history (#1491)
# What it does

Adds pagination to the accounts hold history page.

# Why it is important

This will prevent long page load times for long time users. See #1475. 

# UI Change Screenshot

![Screenshot 2024-05-12 at 3 01
48 AM](https://github.com/chicago-tool-library/circulate/assets/3209502/eb00ed9d-0f97-44c6-b9fe-4c681fa00a14)

# Implementation notes

I wasn't sure how many holds should be shown on each page, so I just
left it as the pagy default.

The test I added might be considered redundant, but I figured I'd leave
it in for now. It technically doesn't test the pagination functionality
itself. I'm happy to make it do so, but I decided to hold off after I
couldn't find a similar test for the other uses of pagy in the app.

I also left a small change in the Gemfile.lock regarding the
installation of Nokogiri on laptops like mine.