Added pagination to the admin users index page under settings (#1496)
# What it does

Adds pagination with pagy to the users index page under settings.

# Why it is important

Currently this page takes just over 2 seconds to load so this should
speed things up significantly. See #1249 for more.

# UI Change Screenshot
![Screenshot 2024-05-16 at 10 55
35 AM](https://github.com/chicago-tool-library/circulate/assets/3209502/ac13834f-d733-4909-84c2-8ca4f3f4f9cd)



# Implementation notes

Just as basic a pagy implementation as I could manage. I also worked on
#1495 which involves filtering this same page. Once one of these is
merged it might take a little work on the other to make sure the filters
and pagination don't cause any issues with one another, which I'm happy
to take care of. Alternatively, these two could be combined? Depends on
how useful one is without the other.