Improve hold text on the item page (#1483)
# What it does

Fixes #1205. Improves the text on the item page related to placing the
item on hold.

* The heading now says "You have this item on hold" if the user has
placed a hold (it used to always say "You can place this item on hold").
* Text is more informative and conditional on whether the user has
placed a hold, the number of users ahead of the logged-in user on the
wait list, and whether the item is ready for pickup.
* The "Place a hold" button only appears if it is needed.
* Added a conditional "Remove Hold" button when the user has already
placed a hold.

# Why it is important

The old text was long and difficult to understand. See #1205 for more
details.

# UI Change Screenshot

The screenshots below show the new UI in a variety of scenarios.

### Logged-in user does not have a hold

Item available, no holds:


![image](https://github.com/chicago-tool-library/circulate/assets/38900370/bfc4e527-2f41-4f71-8940-aff6057f4fdd)

Item available, someone else has a hold:


![image](https://github.com/chicago-tool-library/circulate/assets/38900370/43c5415a-8b21-41d8-a149-07c382c68000)

Item checked out:


![image](https://github.com/chicago-tool-library/circulate/assets/38900370/7a2b3673-1030-44c6-8bd0-bd7574e20877)

### Logged-in user has a hold

Item available, first in line:


![image](https://github.com/chicago-tool-library/circulate/assets/38900370/37c8467a-a097-414f-b3d6-7141eaf9b366)

Item available, first in line, appointment scheduled:


![image](https://github.com/chicago-tool-library/circulate/assets/38900370/943d061b-b1c0-4cbf-9540-f92f80952e32)

Item available, not first in line:


![image](https://github.com/chicago-tool-library/circulate/assets/38900370/223bc7fc-50b1-485e-93ec-595b2a7c5c38)

Item checked out:


![image](https://github.com/chicago-tool-library/circulate/assets/38900370/46ea061d-2e66-47a3-9989-8c4e7edd5838)

### User is not logged in


![image](https://github.com/chicago-tool-library/circulate/assets/38900370/08de1b2f-6312-42c7-8bf5-9b41f0fbb985)

# Implementation notes

I tried to reuse existing functionality where possible (for example, I
used the `render_remove_link` function for the "Remove Hold" button).