Edit reservation holds in place (#1506)
# What it does

Adds the ability to edit the items (`ReservationHolds`) on a reservation
in place in the table. This is likely going to be the way that folks
work with these models as we build out other UIs as well.

Validations are shown on the page properly. There is some UI cleanup
work to do, of course- this is still rough.

# UI Change Screenshot

![2024-05-21 20 09
28](https://github.com/chicago-tool-library/circulate/assets/3331/2915fff2-ce3c-4004-9f21-2efbf3fadd71)

---------

Co-authored-by: Michael Crismali <crismali@users.noreply.github.com>