Staff can create holds or loans for members even if that member has overdue loans (#1544)
# What it does

This allows anyone with staff/admin access to create a hold or a loan
for a member that already has an overdue loan. It removes a check that
prevented this.

# Why it is important

It addresses
[#665](https://github.com/chicago-tool-library/circulate/issues/665) by
removing a roadblock for staff.

# UI Change Screenshot
Before this change, staff would see this toast message regarding the
overdue loan.

![Screenshot 2024-06-13 at 11 26
32 AM](https://github.com/chicago-tool-library/circulate/assets/3209502/2f1290c7-d435-4268-9372-805efd73938a)

After this change, staff will see the usual item "Lend or Hold Item"
form.

![Screenshot 2024-06-13 at 11 26
14 AM](https://github.com/chicago-tool-library/circulate/assets/3209502/6f6032a5-2651-4730-849e-20998c1cb8c9)

# Implementation notes

I just exposed the form in the HTML and didn't have to change any
validations or controller checks.