Prevent codescanning warnings by ensuring methods return constant values and not technically user input (#1548)
# What it does

Refactors some filtering methods to use `case` statements and constants
to so that user input is never technically passed to the `send` method.

# Why it is important

Addresses
[#1507](https://github.com/chicago-tool-library/circulate/issues/1507).
Even if the code was safe before, it's nice to have less noise.

# UI Change Screenshot

Nothing to see here.

# Implementation notes

I added an `INDEX_FILTERS` constant so all of the values are grouped in
an array but nothing uses it. I've seen other patterns for managing
constants and kind of just went with this one, but if there's a
preferred way to do this I'm happy to make those changes.