Create Notifications whenever Devise sends an email (#1543)
# What it does

This creates `Notification`s whenever the Devise mailer is used
(`UserMailer`). It uses the set `@subject` when present and falls back
to the mail object's subject.

# Why it is important

This takes care of
[#1542](https://github.com/chicago-tool-library/circulate/issues/1542)
(this was a draft issue when I started so I converted it into an issue).

# UI Change Screenshot

Nothing much to see here.

# Implementation notes

I used the `@subject` instance variable as the default but fell back to
`mail.subject` because the emails that are not overridden lack
`@subject`.

When I was first looking at this I assumed that Devise or Rails just
automatically picked up on `@subject` and used that as the email's
subject line but that doesn't seem to be the case. The email's actual
subject in this case is "Reset password instructions" and `@subject` is
only used for the `Notification`. Not calling this out because it's
wrong or anything, it just confused me for a moment so I wonder if maybe
there's a better name or if the original intent was for `@subject` to be
the email's subject line.