More attempts to fix InvalidAuthenticityToken issues (#1585)
# What it does

1. Configures AppSignal to track the `USER_AGENT` header so we know if
the issue affect only certain clients. Mobile Safari is a mentioned a
lot in the prior art here, but we don't know if our problems also match
that pattern.
2. Handle InvalidAuthenticityToken errors by redirecting folks back to
where they came from and asking them to try again. In theory this will
help if it turns out that folks are hitting this issue by using a cached
page that's lost its session, although I think the work done in #1478
should have fixed that.

I couldn't figure out how to succinctly trigger an
`InvalidAuthenticityToken` in a controller test, but this code should
only affect folks who have already hit an error anyway.