Switch system tests to use playwright (#1591)
# What it does

Changes from Selenium to Playwright to run our browser based system
tests.

# Why it is important

The hope is that this is a smidge more reliable / less flaky.

# Implementation notes

* This is mostly based on a [recent blog
post](https://justin.searls.co/posts/running-rails-system-tests-with-playwright-instead-of-selenium/)
that reported a good experience switching.
* There were a few places that Playwright had trouble with ambiguous
links - it seems to "see" both the mobile and desktop nav menus. Fixed
these with `within` scopes.
* The drag/drop test needed an extra delay to work... it seemed like it
didn't drag far enough w/o it. I guess playwright is slower to move the
virtual mouse!? 😂
* I decided to yank out Docker support for system tests for now. There
is a [story for setting this back
up](https://playwright-ruby-client.vercel.app/docs/article/guides/playwright_on_alpine_linux),
but it's high effort and nobody is using it right now. We can always
pick it up down the road.
* Cleaned up unused / unnecessary helpers:
   - Richtext selector: included in Capybara / Rails now
   - JS Error detection: no tests fail without it, so seems fine now

---------

Co-authored-by: Jim Benton <jim@autonomousmachine.com>