Added a review notes tab to the reservation show page (#1588)
# What it does

After a reservation is reviewed (approved or rejected) a tab called
"Review Notes" appears that lets admins see the review notes.

# Why it is important

Takes care of #1533 

# UI Change Screenshot

![Screenshot 2024-07-16 at 1 00
50 AM](https://github.com/user-attachments/assets/a6eb9480-ebd8-4044-b929-d2ebb1682fae)

# Implementation notes

My text editor auto-formatted the scss. I can undo that so it's only the
class I added, but since the code and the HTML have their own linters
the changes seem reasonable to me.

I continued the pattern for the other tabs on the reservation show page.
Feels a little heavy to make a full request just for the review notes,
but I think keeping the tab implementations consistent is probably more
important.