Standard rb fixes (#1628)
# What it does

This is just a bunch of standard rb fixes for everything left in the
todo list except for the old migrations.

# Why it is important

Takes us a little closer to being fully standard rb compliant

# Implementation notes

I added a migration for missing uniqueness constraints. I did a smoke
test in blazer to make sure there weren't already records that violated
the constraints but we might want to check those tables out again before
deploying.