Move reservation reviewing into a tab (#1646)
# What it does

Instead of reviewing a reservation on its own page, now it can be
reviewed in a "Review" tab

# Why it is important

Takes care of #1634 

# UI Change Screenshot

Pending reservation:
![Screenshot 2024-08-19 at 4 44
04 PM](https://github.com/user-attachments/assets/6c615736-ff03-4168-a82f-9de278c17f3b)

Requested reservation on the items tab:

![Screenshot_2024-08-19_at_4_44_16 PM](https://github.com/user-attachments/assets/885e8277-22dd-4a0b-a554-9037464c154e)

Requested reservation on the review tab:
![Screenshot 2024-08-19 at 4 44
29 PM](https://github.com/user-attachments/assets/93097ced-44d2-473a-afc9-3376c8f5d3fa)

Immediately after approving a requested reservation:
![Screenshot 2024-08-19 at 4 44
47 PM](https://github.com/user-attachments/assets/1fd1dd16-dc3a-4a67-8a86-a2f4b40bd0f9)

The review notes tab of an approved reservation:
![Screenshot 2024-08-19 at 4 44
57 PM](https://github.com/user-attachments/assets/9d207a42-cba4-4f7d-9121-c1dac17b5835)

# Implementation notes

This is implemented the way the other tabs on the reservation show page
are. I think from an accessibility perspective these tabs are probably
more like a subnav rather than capital-T "Tabs" but that might be a more
general thing to investigate.