Rebuild member navigation menu (#1660)
# What it does

Total rebuild of the main navigation for members on mobile and desktop.

# Why it is important

#1630 

# UI Change Screenshot

Mobile collapsed:
![Screenshot 2024-08-28 at 11 31
35 AM](https://github.com/user-attachments/assets/2c5d924f-ad25-4812-8832-536707b6922f)

Mobile signed in and expanded: 
![Screenshot 2024-08-28 at 11 32
07 AM](https://github.com/user-attachments/assets/06fc6227-92d0-4174-96b6-501a76d66dd5)

Mobile signed out and expanded: 
![Screenshot 2024-08-28 at 11 31
44 AM](https://github.com/user-attachments/assets/988a8166-522f-4131-81c4-4c91785d0784)

Desktop without hover:
![Screenshot 2024-08-28 at 11 30
14 AM](https://github.com/user-attachments/assets/1c70b6f7-cb6c-4841-9466-168f7f3c3764)

Desktop signed in and hovering on Borrow Tools:
![Screenshot 2024-08-28 at 11 29
30 AM](https://github.com/user-attachments/assets/e8058ebb-a11c-4597-ae00-ea893c09c505)

Desktop signed out and hovering on Borrow Tools:
![Screenshot 2024-08-28 at 11 30
21 AM](https://github.com/user-attachments/assets/d529ea2a-8dda-4dce-8acd-4389d7c2b2df)

Desktop signed in as admin and hovering on Membership:
![Screenshot 2024-08-28 at 11 29
39 AM](https://github.com/user-attachments/assets/23a6eb4a-d1a9-4372-999c-d44d181035f0)

Desktop signed out and hovering on Membership:
![Screenshot 2024-08-28 at 11 30
28 AM](https://github.com/user-attachments/assets/d517670a-75f1-46f9-98ea-333a02ca9a84)

Desktop hovering on About Us:
![Screenshot 2024-08-28 at 11 29
50 AM](https://github.com/user-attachments/assets/c930a618-0bfa-4efe-91f2-94c9c7c56e46)

Desktop signed in and hovering on Account:
![Screenshot 2024-08-28 at 11 30
03 AM](https://github.com/user-attachments/assets/c0894097-77d8-4d10-8935-66f9ffa273ad)

# Implementation notes

Pure CSS implementation of the mobile nav sliding over (admin still
depends on the old JS implementation which is why that isn't removed).
The hamburger menu is a kind of secret checkbox label so the styles can
toggle on the `:checked` pseudo class. Clicking off to close it is also
a label for the checkbox, just one with no opacity.

I didn't rely too much on the Spectre CSS classes/components since there
wasn't a good analog for what we wanted on mobile or desktop (members
would have had to click a button to make the dropdown appear instead of
just being able to hover, that sort of thing).

Right now the context menu is really just another hover dropdown since
all it really has is the sign in/out links. "Hours & Locations" will
take members to the account home page since that information seems to
live in the announcements. "Contact Us" and "Visit Our Website" just go
to the tool library site in a new tab. "Tools for Organizations" goes to
My Turn (also in a new tab).

Please give me any kind of feedback, especially in terms of the visual
design 💅 (I'm not a designer so I was just winging it).