Add PendingReservationItem and a few uses of it (#1639)
# What it does

Adds graceful handling of items that aren't a part of a reservation
being scanned while someone is being built. This takes place e.g. the
morning that a reservation is scheduled to be picked up.

We store these items in the database so that they are safely persisted
even if you close/reload the page. As this feature develops further,
there will be a way to update the reservation's holds so that the item
represented by a pending item can be added to the reservation pickup
normally (and the pending item deleted).

This is the first part of #1512.

# Why it is important

If we're scanning hundreds of items, the app needs to persist which ones
were unexpected so the staff member can review them and decide what to
do.

# UI Change Screenshot

<img width="989" alt="Screenshot 2024-08-14 at 11 13 37 PM"
src="https://github.com/user-attachments/assets/4cc5d422-e67f-464a-bbc1-5b54c9d07a30">