Member facing reservation questions (#1674)
# What it does

Members now have to answer any questions in the system when making a
reservation.

# Why it is important

#1520 

# UI Change Screenshot

The new reservation form with a text and number question:
![Screenshot 2024-09-16 at 2 57
21 PM](https://github.com/user-attachments/assets/0e7f3c64-df8b-49c7-92bc-c76a64a51cd6)

The new reservation form with errors:
![Screenshot 2024-09-16 at 2 57
32 PM](https://github.com/user-attachments/assets/8ce45126-3b45-4780-91ed-9daa6b8e9cf5)

The show reservation page with answered questions
![Screenshot 2024-09-16 at 2 58
07 PM](https://github.com/user-attachments/assets/a0a2b553-d683-47af-b239-cdc69ba414ea)

# Implementation notes

Right now a `nil` value for a `number` field is coerced to a `0` so
`number` questions/answers can never really be `blank?`. Not sure if
this is nice or annoying but I'm happy to change it up if that's what
folks prefer.

I didn't get into editing/updating since we're not likely going to have
that as a flow in this part of the app. I thought about removing those
actions/routes/views but I figured that'd likely end up happening once
other ways of interacting with the reservation are worked out.

In order to make the questions optional for admins but required for
members, I tried something a little weird but I think it works well
enough: I subclassed `Answer` into a `RequiredAnswer` that has an
additional validation (same table and everything else though). The risks
for this are if we forget to use `RequiredAnswer` and instead use
`Answer` in member-facing context. I don't think that's particularly
more risky than other implementations of this kind of conditional
validation logic that I've seen, but I'm always open to learning new
approaches!